+++ src/app/snippets/mod.rs
@@ -1,0 +1,4 @@
+mod snippets;
+
+pub(crate) use snippets::Snippet;
+pub(crate) use snippets::standard_postgres_snippets;