+++ src/app/user_event.rs
@@ -6,7 +6,8 @@ use super::{connection::DbResponse, EditorId};
 pub(crate) enum TisqEvent {
     TreeReloaded(SentTree),
     DbResponse(DbResponse),
-    EditorContentReset(EditorId, String), // TODO: use attr instead of UserEvent
+    EditorContentAdd(EditorId, String), // TODO: use attr instead of UserEvent
+    EditorSnippetResolve(EditorId, String), // TODO: use attr instead of UserEvent
 }
 
 // For the purposes of subscriptions we only care about the type of the event
@@ -16,7 +17,8 @@ impl PartialEq for TisqEvent {
         match (self, other) {
             (Self::TreeReloaded(_), Self::TreeReloaded(_)) => true,
             (Self::DbResponse(_), Self::DbResponse(_)) => true,
-            (Self::EditorContentReset(_, _), Self::EditorContentReset(_, _)) => true,
+            (Self::EditorContentAdd(_, _), Self::EditorContentAdd(_, _)) => true,
+            (Self::EditorSnippetResolve(_, _), Self::EditorSnippetResolve(_, _)) => true,
             _ => false,
         }
     }